Fix VIBE_2 to VIBE_4
I was using your codes for use in homeassistant.  VIBE_1 and OFF worked, but 2-4 didn't work.  After packet sniffing, I noticed that the third-to-last hex digit was also changing, whereas you had it as 00 for all 4 commands.  This should fix it.  I tested on my tempur bed to make sure the codes are right.  Please let me know if it now works for you too.